Refactored some (non-locked) methods into game_manager
