+++ src/domain/game.cpp
@@ -47,16 +47,6 @@ std::shared_mutex game_mutex;
 std::list<rps_lobby> lobby_queue;
 
 /**
- * @brief Creating rps_bot
- */
-dpp::cluster *creator{nullptr};
-
-void init(dpp::cluster &bot) {
-  creator = &bot;
-  creator->log(dpp::ll_info, "Game state initialized");
-}
-
-/**
  * @brief PROTECTED
  *
  * @param player_id
@@ -240,12 +230,20 @@ rps_lobby get_lobby(const unsigned int lobby_id) {
   return {};
 }
 
+/**
+ * @brief Get the player score object
+ *
+ * PROTECTED
+ * @param lobby_id
+ * @param index
+ * @return unsigned int
+ */
 unsigned int get_player_score(const unsigned int lobby_id,
                               const unsigned int index) {
   std::lock_guard<std::shared_mutex> game_lock(game_mutex);
   for (const auto &lobby : lobby_queue) {
     if (lobby.id == lobby_id) {
-      return lobby.players.at(index)->score;
+      return lobby.players[index]->score;
     }
   }
   return 0;
@@ -263,7 +261,7 @@ std::shared_ptr<player_info> get_player_info(const unsigned int lobby_id,
   std::lock_guard<std::shared_mutex> game_lock(game_mutex);
   for (const auto &lobby : lobby_queue) {
     if (lobby.id == lobby_id) {
-      return lobby.players.at(index);
+      return lobby.players[index];
     }
   }
   return nullptr;
@@ -282,7 +280,7 @@ dpp::snowflake get_player_id(const unsigned int lobby_id,
   std::lock_guard<std::shared_mutex> game_lock(game_mutex);
   for (const auto &lobby : lobby_queue) {
     if (lobby.id == lobby_id) {
-      return lobby.players.at(player_index)->player.id;
+      return lobby.players[player_index]->player.id;
     }
   }
   return 0;
@@ -444,41 +442,6 @@ std::string get_player_name(const unsigned int lobby_id,
   return "";
 }
 
-void send_game_messages(const unsigned int lobby_id) {
-  rps_lobby found_lobby = get_lobby(lobby_id);
-  if (found_lobby.id == 0) {
-    creator->log(dpp::ll_critical, "Could not find lobby");
-    return;
-  }
-  unsigned int game_num = get_game_num(lobby_id);
-
-  if (game_num == 1) {
-    for (auto &player_info : found_lobby.players) {
-      clear_queue_timer(player_info->player.id);
-    }
-  }
-
-  std::string player_one_name = get_player_name(lobby_id, 0);
-  unsigned int player_one_score = get_player_score(lobby_id, 0);
-  std::string player_two_name = get_player_name(lobby_id, 1);
-  unsigned int player_two_score = get_player_score(lobby_id, 1);
-
-  start_game_timer(lobby_id, creator->start_timer(
-                                 [=](unsigned long t) {
-                                   handle_timeout(lobby_id);
-                                   creator->stop_timer(t);
-                                 },
-                                 GAME_TIMEOUT));
-
-  for (const auto &player_info : found_lobby.players) {
-    creator->direct_message_create(
-        player_info->player.id,
-        embeds::game(player_info->init_interaction, lobby_id, game_num,
-                     player_one_name, player_one_score, player_two_name,
-                     player_two_score));
-  }
-}
-
 bool is_game_complete(const unsigned int lobby_id) {
   std::lock_guard<std::shared_mutex> game_lock(game_mutex);
   for (const auto &lobby : lobby_queue) {
@@ -488,113 +451,6 @@ bool is_game_complete(const unsigned int lobby_id) {
     }
   }
   return false;
-}
-
-void send_result_messages(const unsigned int lobby_id,
-                          const unsigned int winner, const unsigned int loser,
-                          bool draw) {
-  unsigned int game_num = get_game_num(lobby_id);
-  std::string player_one_name = get_player_name(lobby_id, 0);
-  std::string player_one_choice = get_player_choice(get_player_id(lobby_id, 0));
-  unsigned int player_one_score = get_player_score(lobby_id, 0);
-  std::string player_two_name = get_player_name(lobby_id, 1);
-  std::string player_two_choice = get_player_choice(get_player_id(lobby_id, 1));
-  unsigned int player_two_score = get_player_score(lobby_id, 1);
-  dpp::slashcommand_t winner_int = get_player_interaction(lobby_id, winner);
-  dpp::slashcommand_t loser_int = get_player_interaction(lobby_id, loser);
-
-  dpp::message msg_win = embeds::game_result(
-      winner_int, game_num, player_one_name, player_one_choice, player_two_name,
-      player_two_choice, "WIN");
-  dpp::message msg_loss = embeds::game_result(
-      loser_int, game_num, player_one_name, player_one_choice, player_two_name,
-      player_two_choice, "LOSS");
-  if (draw) {
-    msg_win = embeds::game_result(winner_int, game_num, player_one_name,
-                                  player_one_choice, player_two_name,
-                                  player_two_choice, "DRAW");
-    msg_loss = embeds::game_result(loser_int, game_num, player_one_name,
-                                   player_one_choice, player_two_name,
-                                   player_two_choice, "DRAW");
-  }
-
-  /* These need to be sent before the next game message is sent, so we make them
-   * synchronous */
-  creator->direct_message_create_sync(get_player_id(lobby_id, winner), msg_win);
-  creator->direct_message_create_sync(get_player_id(lobby_id, loser), msg_loss);
-
-  /* Hack to set player emoji */
-  std::string player_one_emoji_choice =
-      (player_one_choice == "Rock")
-          ? ":rock:"
-          : ((player_one_choice == "Paper") ? ":page_facing_up:"
-                                            : ":scissors:");
-  std::string player_two_emoji_choice =
-      (player_two_choice == "Rock")
-          ? ":rock:"
-          : ((player_two_choice == "Paper") ? ":page_facing_up:"
-                                            : ":scissors:");
-
-  /* Create normal text message for result (may have a higher rate limit?) */
-  dpp::message result_msg;
-  if (draw) {
-    result_msg = dpp::message(fmt::format(
-        "__**Lobby #{} - Game {}**__\n{}  {}  {}  |  {}  {}  {}", lobby_id,
-        game_num, player_one_name, player_one_emoji_choice, player_one_score,
-        player_two_score, player_two_emoji_choice, player_two_name));
-  } else {
-    /* Determine which name + score to bold */
-    if (winner == 0) {
-      result_msg = dpp::message(fmt::format(
-          "__**Lobby #{} - Game {}**__\n**{}**  {}  **{}**  |  {}  {}  {}",
-          lobby_id, game_num, player_one_name, player_one_emoji_choice,
-          player_one_score, player_two_score, player_two_emoji_choice,
-          player_two_name));
-    } else {
-      result_msg = dpp::message(fmt::format(
-          "__**Lobby #{} - Game {}**__\n{}  {}  {}  |  **{}**  {}  **{}**",
-          lobby_id, game_num, player_one_name, player_one_emoji_choice,
-          player_one_score, player_two_score, player_two_emoji_choice,
-          player_two_name));
-    }
-  }
-
-  /* Send results in channels that players queued in */
-  dpp::slashcommand_t player_one_interaction =
-      get_player_interaction(lobby_id, 0);
-  dpp::slashcommand_t player_two_interaction =
-      get_player_interaction(lobby_id, 1);
-
-  /* Both of them were in DMs, so no work needed */
-  if ((player_one_interaction.command.guild_id.empty() ||
-       player_one_interaction.command.guild_id == 0) &&
-      (player_two_interaction.command.guild_id.empty() ||
-       player_two_interaction.command.guild_id == 0)) {
-    return;
-  }
-
-  /* Just send one if they are the same */
-  if (player_one_interaction.command.channel_id ==
-      player_two_interaction.command.channel_id) {
-    creator->message_create(
-        result_msg.set_guild_id(player_one_interaction.command.guild_id)
-            .set_channel_id(player_one_interaction.command.channel_id));
-    return;
-  }
-
-  if (!player_one_interaction.command.guild_id.empty() &&
-      player_one_interaction.command.guild_id != 0) {
-    creator->message_create(
-        result_msg.set_guild_id(player_one_interaction.command.guild_id)
-            .set_channel_id(player_one_interaction.command.channel_id));
-  }
-
-  if (!player_two_interaction.command.guild_id.empty() &&
-      player_two_interaction.command.guild_id != 0) {
-    creator->message_create(
-        result_msg.set_guild_id(player_two_interaction.command.guild_id)
-            .set_channel_id(player_two_interaction.command.message_id));
-  }
 }
 
 /**
@@ -639,7 +495,7 @@ void start_queue_timer(const dpp::snowflake player_id, dpp::timer timer) {
  *
  * @param player_id
  */
-void clear_queue_timer(const dpp::snowflake player_id) {
+void clear_queue_timer(dpp::cluster *creator, const dpp::snowflake player_id) {
   std::lock_guard<std::shared_mutex> game_lock(game_mutex);
   for (auto &lobby : lobby_queue) {
     for (auto &player_info : lobby.players) {
@@ -672,7 +528,7 @@ void start_game_timer(const unsigned int lobby_id, dpp::timer timer) {
  *
  * @param lobby_id
  */
-void clear_game_timer(const unsigned int lobby_id) {
+void clear_game_timer(dpp::cluster *creator, const unsigned int lobby_id) {
   std::lock_guard<std::shared_mutex> game_lock(game_mutex);
   for (auto &lobby : lobby_queue) {
     if (lobby.id == lobby_id) {
@@ -680,133 +536,6 @@ void clear_game_timer(const unsigned int lobby_id) {
       return;
     }
   }
-}
-
-void send_match_results(const unsigned int lobby_id, const dpp::user &winner,
-                        bool double_afk = false) {
-  dpp::slashcommand_t player_one_interaction =
-      get_player_interaction(lobby_id, 0);
-  dpp::slashcommand_t player_two_interaction =
-      get_player_interaction(lobby_id, 1);
-
-  dpp::message player_one_message = embeds::match_result(
-      player_one_interaction, lobby_id, get_game_num(lobby_id),
-      get_player_name(lobby_id, 0), get_player_score(lobby_id, 0),
-      get_player_name(lobby_id, 1), get_player_score(lobby_id, 1), winner,
-      double_afk);
-
-  dpp::message player_two_message = embeds::match_result(
-      player_two_interaction, lobby_id, get_game_num(lobby_id),
-      get_player_name(lobby_id, 0), get_player_score(lobby_id, 0),
-      get_player_name(lobby_id, 1), get_player_score(lobby_id, 1), winner,
-      double_afk);
-
-  creator->direct_message_create(get_player_id(lobby_id, 0),
-                                 player_one_message);
-  creator->direct_message_create(get_player_id(lobby_id, 1),
-                                 player_two_message);
-
-  /* Send results in channels that players queued in */
-  dpp::message msg = embeds::match_result(
-      dpp::interaction_create_t(), lobby_id, get_game_num(lobby_id),
-      get_player_name(lobby_id, 0), get_player_score(lobby_id, 0),
-      get_player_name(lobby_id, 1), get_player_score(lobby_id, 1), winner,
-      double_afk);
-
-  /* Both of them were in DMs, so no work needed */
-  if ((player_one_interaction.command.guild_id.empty() ||
-       player_one_interaction.command.guild_id == 0) &&
-      (player_two_interaction.command.guild_id.empty() ||
-       player_two_interaction.command.guild_id == 0)) {
-    return;
-  }
-
-  /* Just send one if they are the same */
-  if (player_one_interaction.command.channel_id ==
-      player_two_interaction.command.channel_id) {
-    creator->message_create(
-        msg.set_guild_id(player_one_interaction.command.guild_id)
-            .set_channel_id(player_one_interaction.command.channel_id));
-    return;
-  }
-
-  if (!player_one_interaction.command.guild_id.empty() &&
-      player_one_interaction.command.guild_id != 0) {
-    creator->message_create(
-        msg.set_guild_id(player_one_interaction.command.guild_id)
-            .set_channel_id(player_one_interaction.command.channel_id));
-    return;
-  }
-
-  if (!player_two_interaction.command.guild_id.empty() &&
-      player_two_interaction.command.guild_id != 0) {
-    creator->message_create(
-        msg.set_guild_id(player_two_interaction.command.guild_id)
-            .set_channel_id(player_two_interaction.command.message_id));
-    return;
-  }
-}
-
-void handle_choice(const dpp::button_click_t &event) {
-  /* Find player lobby */
-  unsigned int player_lobby_id =
-      find_player_lobby_id(event.command.get_issuing_user().id);
-
-  rps_lobby player_lobby = get_lobby(player_lobby_id);
-
-  /* 1. Go set the choice, then send a confirmation message */
-  clear_queue_timer(event.command.get_issuing_user().id);
-  set_player_choice(event.command.get_issuing_user().id, event.custom_id);
-  creator->direct_message_create(
-      event.command.get_issuing_user().id,
-      dpp::message(fmt::format("You selected {}! {}", event.custom_id,
-                               tr("E_WAITING", event))));
-
-  /* 2. If both choices are selected, determine who won and increment winner
-   */
-  if (check_both_responses(player_lobby_id)) {
-    clear_game_timer(player_lobby_id);
-    std::string result = determine_winner(player_lobby_id);
-    dpp::user winner;
-    if (result == "1") {
-      winner = get_player_info(player_lobby_id, 0)->player;
-      increment_player_score(player_lobby_id, 0);
-      send_result_messages(player_lobby_id, 0, 1);
-    } else if (result == "2") {
-      winner = get_player_info(player_lobby_id, 1)->player;
-      increment_player_score(player_lobby_id, 1);
-      send_result_messages(player_lobby_id, 1, 0);
-    } else if (result == "D") {
-      send_result_messages(player_lobby_id, 0, 1, true);
-    }
-
-    if (is_game_complete(player_lobby_id)) {
-      /* Finish up */
-      send_match_results(player_lobby_id, winner);
-      remove_lobby_from_queue(player_lobby_id, true);
-    } else {
-      /* Send message */
-      increment_game_num(player_lobby_id);
-      reset_choices(player_lobby_id);
-      send_game_messages(player_lobby_id);
-    }
-  }
-}
-
-void handle_timeout(const unsigned int lobby_id) {
-  std::string result = determine_winner(lobby_id);
-  dpp::user winner;
-  if (result == "1") {
-    increment_player_score(lobby_id, 0);
-    send_match_results(lobby_id, get_player_info(lobby_id, 0)->player);
-  } else if (result == "2") {
-    increment_player_score(lobby_id, 1);
-    send_match_results(lobby_id, get_player_info(lobby_id, 1)->player);
-  } else if (result == "FF") {
-    send_match_results(lobby_id, winner, true);
-  }
-
-  remove_lobby_from_queue(lobby_id, true);
 }
 
 } // namespace game
